+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 2023-10-04
+
 ### Added
 
 - The ability to set custom despawn and component removal functions.
@@ -186,7 +188,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release after separation from [lifescape](https://github.com/lifescapegame/lifescape).
 
-[unreleased]: https://github.com/lifescapegame/bevy_replicon/compare/v0.12.0...HEAD
+[unreleased]: https://github.com/lifescapegame/bevy_replicon/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.9.1...v0.10.0